MerkleTree (merkle_tree v0.1.0)

(Sparse) Merkle Tree.

MerkleTree implements Enumerable and Collectable protocols.

key()

@type key() :: term()

A tree key.

opts()

@type opts() :: [{:auto_update_hash, boolean()}]

MerkleTree options:

auto_update_hash: tree hashes are updated on every put/3 / delete/2. If set to false hashes should be explicitly computed calling update_hashes/1 before calling functions like merkle_equal?/2. Defaults to true.

t()

@type t() :: %MerkleTree{opts: opts(), size: non_neg_integer(), tree: tree()}

Merkle tree.

value()

@type value() :: term()

A tree value.

delete(merkle_tree, key)

@spec delete(t(), key()) :: t()

Deletes a tree key.

equal?(mt1, mt2)

@spec equal?(t(), t()) :: boolean()

Checks if two trees are equal.

The equality is first checked leveraging the merkle tree hashes and in case of uncertaintainty with a default comparison algorithm.

Raises if the hashes are not calculated (Refer to :auto_update_hash opts/0).

See also merkle_equal?/2.

fetch(merkle_tree, key)

@spec fetch(t(), key()) :: {:ok, value()} | :error

Fetches a key value from the tree.

has_key?(merkle_tree, key)

@spec has_key?(t(), key()) :: boolean()

Returns whether the given key exists in the tree.

keys(merkle_tree)

@spec keys(t()) :: [key()]

Returns all keys in the tree.

merkle_diff(merkle_tree1, merkle_tree2)

@spec merkle_diff(t(), t()) :: [key()]

Returns the keys that differ (changed / removed / added) between two trees.

The diff is computed leveraging the merkle tree hashes and hence in case of hash collisions a wrong result could be returned.

More preciselly the diff reported is correct but could be incomplete.

Raises if the hashes are not calculated (Refer to :auto_update_hash opts/0).

merkle_equal?(merkle_tree1, merkle_tree2)

@spec merkle_equal?(t(), t()) :: boolean()

Checks if two trees are (merkle) equal.

The equality is checked leveraging the merkle tree hashes and hence in case of hash collisions a wrong result could be returned.

More precisely if it reports that the trees are different than this is 100% true, otherwise it means the trees are "probably" equal.

Raises if the hashes are not calculated (Refer to :auto_update_hash opts/0).

See also equal?/2.

new(enum \\ nil, opts \\ [])

@spec new(Enumerable.t({key(), value()}) | nil, opts()) :: t()

Create a new MerkleTree.t/0 from an optional enumerable of key->value pairs.

If enumerable is not provided it creates an empty tree. The opts optional parameter can be provided to customize the tree behaviour (refer to t:opts/1).

put(merkle_tree, key, value)

@spec put(t(), key(), value()) :: t()

Puts the given key/value in the tree.

size(merkle_tree)

@spec size(t()) :: non_neg_integer()

Returns the number of keys stored in the tree.

update_hashes(mt)

@spec update_hashes(t()) :: t()

Update the tree hashes.

Refer to :auto_update_hash opts/0.