MerkleTree (merkle_tree v0.1.0)

MerkleTree

(Sparse) Merkle Tree.

MerkleTree implements Enumerable and Collectable protocols.

diff_act()

@type diff_act() :: :add | :del | :upd

diff_op()

@type diff_op() :: {diff_act(), key()}

item()

@type item() :: {key(), value()}

A tree item - a key/value pair.

iterator()
(opaque)

@opaque iterator()

A tree iterator.

key()

@type key() :: term()

A tree key.

opts()

@type opts() :: [{:auto_comp_hash, boolean()}]

MerkleTree options.

auto_comp_hash: tree hashes are computed on every put/3 / delete/2. If set to false hashes should be explicitly computed calling compute_hashes/1 before calling functions like merkle_equal?/2. Defaults to true.

t()

@type t() :: %MerkleTree{opts: opts(), size: non_neg_integer(), tree: tree()}

Merkle tree.

value()

@type value() :: term()

A tree value.

compute_hashes(mt)

@spec compute_hashes(t()) :: t()

Compute the tree hashes.

Refer to :auto_comp_hash opts/0.

delete(merkle_tree, key)

@spec delete(t(), key()) :: t()

Deletes a tree key.

fetch(merkle_tree, key)

@spec fetch(t(), key()) :: {:ok, value()} | :error

Fetches a key value from the tree.

has_key?(mt, key)

@spec has_key?(t(), key()) :: boolean()

Returns whether the given key exists in the tree.

keys(merkle_tree)

@spec keys(t()) :: [key()]

Returns all keys in the tree.

new(enum \\ nil, opts \\ [])

@spec new(Enumerable.t({key(), value()}) | nil, opts()) :: t()

Create a new MerkleTree.t/0 from an optional enumerable of key->value pairs.

If enumerable is not provided it creates an empty tree. The opts optional parameter can be provided to customize the tree behaviour (refer to t:opts/1).

put(merkle_tree, key, value)

@spec put(t(), key(), value()) :: t()

Puts the given key/value in the tree.

size(merkle_tree)

@spec size(t()) :: non_neg_integer()

Returns the number of keys stored in the tree.

first(merkle_tree)

@spec first(t()) :: item() | nil

Returns the first {key, value} according to the merkle tree order.

iterator(mt)

@spec iterator(t()) :: iterator()

Returns an iterator over the tree items, the iteration follows the merkle key tree order.

The iteration algorithm doesn't need to build the complete list of all elements at one time.

last(merkle_tree)

@spec last(t()) :: item() | nil

Returns the last {key, value} according to the merkle tree order.

next(arg)

@spec next(iterator()) :: {item(), iterator()} | nil

Returns the next key in the iteration.

See also iterator/1.

diff(mt1, mt2)

@spec diff(from :: t(), to :: t()) :: [diff_op()]

Returns the diff between two trees.

The diff is expressed as add/delete/update operation to apply to the first tree to get to the second one.

The diff is NOT computed leveraging the merkle tree hashes.

merkle_diff(merkle_tree1, merkle_tree2)

@spec merkle_diff(from :: t(), to :: t()) :: [diff_op()]

Returns the diff between two trees.

The diff is expressed as add/delete/update operation to apply to the first tree to get to the second one.

The diff is computed leveraging the merkle tree hashes and hence in case of hash collisions a wrong result could be returned. More precisely the diff reported is correct but could be incomplete.

Raises if the hashes are not calculated (Refer to :auto_comp_hash opts/0).

equal?(mt1, mt2)

@spec equal?(t(), t()) :: boolean()

Checks if two trees are equal.

The equality is first checked leveraging the merkle tree hashes and in case of uncertainty with a default element by element comparison.

Raises if the hashes are not calculated (Refer to :auto_comp_hash opts/0).

See also merkle_equal?/2.

merkle_equal?(merkle_tree1, merkle_tree2)

@spec merkle_equal?(t(), t()) :: boolean()

Checks if two trees are (merkle) equal.

The equality is checked leveraging the merkle tree hashes and hence in case of hash collisions a wrong result could be returned. More precisely if it reports that the trees are different than this is 100% true, otherwise it means the trees are "probably" equal.

Raises if the hashes are not calculated (Refer to :auto_comp_hash opts/0).

See also equal?/2.